+++ shoop/admin/modules/products/views/variation/__init__.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-# This file is part of Shoop.
-#
-# Copyright (c) 2012-2016, Shoop Ltd. All rights reserved.
-#
-# This source code is licensed under the AGPLv3 license found in the
-# LICENSE file in the root directory of this source tree.
-
-from .edit_variation import ProductVariationView  # noqa